+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.6.5] - Unreleased
+### Added
+- KiCad 8 support
+- Panelize: support for all new options (upto 1.5.1)
+
 ### Changed
 - KiRi: continue even on corrupted schematics (#583)
 - Variants: avoid W045 on nameless pads. Assuming they are on purpose and not